requirements.txt
fastapi
python-dotenv
uvicorn
requests
PyJWT

